Add a page that lists films
This looks nice, but would need to download a lot of large images for no reason;
can be done on the server side instead in the final product.

Rename old "film" page to avoid confusion.

Keep css style files separate for now while a lot of experimentation is going on.