Voting functionality for the new-year smackdown
* Parameterise html template file in generate.py.
The default value remains the standard film list page,
which is most commonly used.

* Add a voting script that brings films up/down or removes them

* Add js folder for it

* Create new template for a film list page with voting functionality
Navigation has been removed for this type of page,
as voting does not support a multi-page spread of films.
Note that leaving it in creates visible empty space on appending elements with vote.js.
Not a good long-term solution as
1) it would silently misbehave if the number of movies on the voting page
exceeds the movies_per_page parameter.
2) drives functionality further away from the main template,
which will encourage duplication.

* Expand css style to cover voting buttons.
This may be better off in a separate file

* Use external icons from unpkg.com